requirements.txt
Flask
requests
pytest
python-dotenv
flask-sqlalchemy